requirements.txt
mock
nose
tabulate
Flask
Flask-SQLAlchemy
numpy
